searcher: move search into an internal pkg
This code is not meant to be imported by other packages. Lets prevent
accidental imports by moving it into internal.

This is a mechanical commit. I just ran "git mv search internal/search".
I then had to fix only two import paths.

Test Plan: CI
