+++ cmd/searcher/internal/search/filter.go
