+++ cmd/searcher/internal/search/filter_test.go
