+++ cmd/searcher/internal/search/search_regex.go
