+++ cmd/searcher/internal/search/search_regex_test.go
