+++ cmd/searcher/internal/search/search_structural.go
