+++ cmd/searcher/internal/search/search_structural_test.go
