+++ cmd/searcher/internal/search/store_test.go
