+++ cmd/searcher/internal/search/zip_test.go
