+++ cmd/searcher/internal/search/zipcache.go
