+++ cmd/searcher/internal/search/zipcache_test.go
